feat: typescript support (#9)
