+++ packages/import-sort-style-4catalyzer/parser-ts.js
@@ -1,0 +1,3 @@
+const createParser = require('./create-parser');
+
+module.exports = createParser({ typescript: true });
